map: puts together layer list and legend
