+++ frontend/src/components/map/index.tsx
@@ -144,8 +144,6 @@ export const Map: FC<CustomMapProps> = ({
 
 export default Map;
 export { default as ZoomControls } from './zoom-controls';
-export { default as LayersDropdown } from './layers-dropdown';
-export { default as Legend } from './legend';
 export { default as Attributions } from './attributions';
 export { default as DrawControls } from './draw-controls';
 export { default as Drawing } from './drawing';
