+++ frontend/src/containers/data-tool/content/map/layers-toolbox/legend/items.tsx
