+++ frontend/src/styles/fonts/figtree/Figtree-VariableFont_wght.ttf
